// apps/frontend/src/api/generated/jobFilterDtoResponse.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Job Research AI
 * Find job faster with Job Research AI
 * OpenAPI spec version: 1.0.0
 */

export interface JobFilterDtoResponse {
  /** List of unique job positions/titles available */
  positions: string[];
  /** List of unique locations available */
  locations: string[];
  /** List of derived job boards or services */
  services: string[];
  /** The minimum salary value found in the current dataset */
  salaryFrom: number;
  /** The maximum salary value found in the current dataset */
  salaryTo: number;
}